requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
databases==0.7.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
pydantic==1.10.4
PyJWT==2.6.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
